Flytt klientoppsettet for tokenveksling til commons
Testen bygget maskinportengeneratoren og klienten inne i selve testfunksjonen. Nå ligger de i mappas commons sammen med setup, slik som de andre mappene gjør det.

Refs #432
